Misc. code changes
Changes based on enabling -Weverything (all warnings)

• Removed few unused macros
• Replaced some argument types with “id” because of type mismatch
(table view delegates)
• Modified TXHumanReadableTimeInterval() to use NSInvocation instead of
objc_msgSend because the return value is an NSInteger
• Fixed some numbers being cast as floats when they are supposed to be
doubles
• Fixed IRCChannelConfig returning self.secretValue (keychain value)
when its supposed to return _secretKey (temporary value in memory)
• Fixed [IRCUser presentAwayMessageFor301] performing incorrect
comparison

+++ Classes/Headers/IRCWorld.h
@@ -52,7 +52,7 @@ TEXTUAL_EXTERN NSString * const IRCWorldClientListWasModifiedNotification;
 @property (nonatomic, assign) TXUnsignedLongLong bandwidthOut;
 @property (nonatomic, assign) BOOL isPopulatingSeeds;
 @property (nonatomic, copy) NSArray *clientList; // clientList as a proxy setter/getter for the internal storage.
-@property (nonatomic, assign) float textSizeMultiplier;
+@property (nonatomic, assign) double textSizeMultiplier;
 
 - (void)setupConfiguration;
 - (void)setupOtherServices;
